+++ bliss/core/scan.cpp
@@ -353,7 +353,7 @@ void bliss::scan::push_device() {
     }
 }
 
-bliss::scan bliss::scan::slice_scan_channels(int start_channel, int count) {
+bliss::scan bliss::scan::slice_scan_channels(int64_t start_channel, int64_t count) {
     if (count == -1) {
         fmt::print("INFO: Got count of -1 channels, automatically extending to last coarse channel ({})\n", get_number_coarse_channels(), get_number_coarse_channels() - start_channel);
         count = get_number_coarse_channels() - start_channel;
